% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{obs_input}
\alias{obs_input}
\title{A data frame of observed spike in ERCC normalized LCPM data}
\format{
A data frame with 92 rows and 50 Columns: Each row is an ERCC
    transcript, and each column is a sample.  Data are read
    coverage-normalized LCPM.
\describe{
  \item{X}{This first column is ERCC spike in mRNA Ids (ERCC-00002 -- ERCC-00171)}
  \item{SAM15}{Sample 15 containing normalized log2 counts per million}
  \item{SAM36}{Sample 36 containing normalized log2 counts per million}
  \item{SAM19}{Sample 19 containing normalized log2 counts per million}
  \item{SAM53}{Sample 53 containing normalized log2 counts per million}
  \item{SAM42}{Sample 42 containing normalized log2 counts per million}
  \item{SAM32}{Sample 32 containing normalized log2 counts per million}
  \item{SAM18}{Sample 18 containing normalized log2 counts per million}
  \item{SAM48}{Sample 48 containing normalized log2 counts per million}
  \item{SAM26}{Sample 26 containing normalized log2 counts per million}
  \item{SAM37}{Sample 37 containing normalized log2 counts per million}
  \item{SAM38}{Sample 38 containing normalized log2 counts per million}
  \item{SAM29}{Sample 29 containing normalized log2 counts per million}
  \item{SAM17}{Sample 17 containing normalized log2 counts per million}
  \item{SAM41}{Sample 41 containing normalized log2 counts per million}
  \item{SAM09}{Sample 09 containing normalized log2 counts per million}
  \item{SAM07}{Sample 07 containing normalized log2 counts per million}
  \item{SAM14}{Sample 14 containing normalized log2 counts per million}
  \item{SAM02}{Sample 02 containing normalized log2 counts per million}
  \item{SAM05}{Sample 05 containing normalized log2 counts per million}
  \item{SAM25}{Sample 25 containing normalized log2 counts per million}
  \item{SAM08}{Sample 08 containing normalized log2 counts per million}
  \item{SAM28}{Sample 28 containing normalized log2 counts per million}
  \item{SAM44}{Sample 44 containing normalized log2 counts per million}
  \item{SAM04}{Sample 04 containing normalized log2 counts per million}
  \item{SAM10}{Sample 10 containing normalized log2 counts per million}
  \item{SAM31}{Sample 31 containing normalized log2 counts per million}
  \item{SAM21}{Sample 21 containing normalized log2 counts per million}
  \item{SAM20}{Sample 20 containing normalized log2 counts per million}
  \item{SAM52}{Sample 52 containing normalized log2 counts per million}
  \item{SAM46}{Sample 46 containing normalized log2 counts per million}
  \item{SAM01}{Sample 01 containing normalized log2 counts per million}
  \item{SAM13}{Sample 13 containing normalized log2 counts per million}
  \item{SAM39}{Sample 39 containing normalized log2 counts per million}
  \item{SAM49}{Sample 49 containing normalized log2 counts per million}
  \item{SAM30}{Sample 30 containing normalized log2 counts per million}
  \item{SAM50}{Sample 50 containing normalized log2 counts per million}
  \item{SAM11}{Sample 11 containing normalized log2 counts per million}
  \item{SAM35}{Sample 35 containing normalized log2 counts per million}
  \item{SAM06}{Sample 06 containing normalized log2 counts per million}
  \item{SAM27}{Sample 27 containing normalized log2 counts per million}
  \item{SAM33}{Sample 33 containing normalized log2 counts per million}
  \item{SAM22}{Sample 22 containing normalized log2 counts per million}
  \item{SAM24}{Sample 24 containing normalized log2 counts per million}
  \item{SAM16}{Sample 16 containing normalized log2 counts per million}
  \item{SAM34}{Sample 34 containing normalized log2 counts per million}
  \item{SAM03}{Sample 03 containing normalized log2 counts per million}
  \item{SAM47}{Sample 47 containing normalized log2 counts per million}
  \item{SAM40}{Sample 40 containing normalized log2 counts per million}
  \item{SAM23}{Sample 23 containing normalized log2 counts per million}
}
}
\usage{
obs_input
}
\description{
A data frame of observed gene expression results for ERCC RNA that was spiked
into samples.  Data are read coverage-normalized log2 counts per million (LCPM).
}
\keyword{datasets}
