% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLowLcpmCutoff.R
\name{print.empLCPM}
\alias{print.empLCPM}
\title{Print the empirically derived LCPM cutoff results}
\usage{
\method{print}{empLCPM}(x, ...)
}
\arguments{
\item{x}{A 'empLCPM' object from \code{\link{getLowLcpmCutoff}}.}

\item{...}{Additional arguments are ignored.}
}
\value{
The function \code{print.empLCPM} extracts the key results and
returns a string of the summary.
}
\description{
Print method for class "empLCPM".
}
