% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLowLcpmCutoff.R
\name{summary.empLCPM}
\alias{summary.empLCPM}
\title{Summarizing the empirically derived LCPM cutoff}
\usage{
\method{summary}{empLCPM}(object, ...)
}
\arguments{
\item{object}{A 'empLCPM' object from \code{\link{getLowLcpmCutoff}}.}

\item{...}{Additional arguments are ignored.}
}
\value{
The function \code{summary.empLCPM} extracts the key results and
returns a string of the summary.
}
\description{
Summary method for class "empLCPM".
}
